v4l2: Various cross-compiling fixes
Tested using the following triplets:

 * armv7-unknown-linux-gnueabihf
 * armv7-unknown-linux-musleabihf
 * x86_64-unknown-linux-gnu
 * x86_64-unknown-linux-musl

Both the default features (v4l2-sys) and the libv4l bindings
(--no-default-features --features v4l-sys) were tested.

+++ src/v4l2/vidioc.rs
@@ -5,7 +5,7 @@ use crate::v4l_sys::*;
 pub type _IOC_TYPE = std::os::raw::c_ulong;
 #[cfg(target_env = "musl")]
 #[allow(non_camel_case_types)]
-pub type _IOC_TYPE = std::os::raw::c_long;
+pub type _IOC_TYPE = std::os::raw::c_int;
 
 // linux ioctl.h
 const _IOC_NRBITS: u8 = 8;
